requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
decorators==2.0.7
defusedxml==0.7.1
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
Faker==15.3.4
greenlet==2.0.1
httpie==3.2.1
idna==3.4
kombu==5.2.4
mimesis==6.1.1
multidict==6.0.2
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.33
psycopg2==2.9.5
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
qrcode==7.3.1
redis==4.3.5
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
